+++ src/blocking/prelude.rs
@@ -1,0 +1,2 @@
+#[doc(no_inline)]
+pub use crate::{blocking::*, prelude::*};
