+++ src/internal/user_api/mod.rs
@@ -102,6 +102,29 @@ impl UserCreateResult {
     }
 }
 
+/// Public/Private asymmetric keypair that is used for decryption/encryption.
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+pub struct KeyPair {
+    public_key: PublicKey,
+    private_key: PrivateKey,
+}
+impl KeyPair {
+    pub fn new(public_key: RecryptPublicKey, private_key: RecryptPrivateKey) -> Self {
+        KeyPair {
+            public_key: public_key.into(),
+            private_key: private_key.into(),
+        }
+    }
+
+    pub fn public_key(&self) -> &PublicKey {
+        &self.public_key
+    }
+
+    pub fn private_key(&self) -> &PrivateKey {
+        &self.private_key
+    }
+}
+
 /// Bundle of information for adding a device
 pub struct DeviceAdd {
     /// public key of the user
@@ -318,6 +341,64 @@ pub async fn user_rotate_private_key<CR: rand::CryptoRng + rand::RngCore>(
     )
     .await?
     .into())
+}
+
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+pub struct DeviceAddResult {
+    /// The user's given id, which uniquely identifies them inside the segment.
+    account_id: UserId,
+    /// The user's segment id
+    segment_id: usize,
+    /// The private key which was generated for a particular device for the user. Not the user's master private key.
+    device_private_key: PrivateKey,
+    /// The signing key which was generated for the device. “expanded private key” (both pub/priv)
+    signing_private_key: DeviceSigningKeyPair,
+    /// The id of the device that was added
+    device_id: DeviceId,
+    /// The name of the device that was added
+    name: Option<DeviceName>,
+    /// The date and time that the device was created
+    created: DateTime<Utc>,
+    /// The date and time that the device was last updated
+    last_updated: DateTime<Utc>,
+}
+
+impl DeviceAddResult {
+    pub fn account_id(&self) -> &UserId {
+        &self.account_id
+    }
+    pub fn segment_id(&self) -> usize {
+        self.segment_id
+    }
+    pub fn signing_private_key(&self) -> &DeviceSigningKeyPair {
+        &self.signing_private_key
+    }
+    pub fn device_private_key(&self) -> &PrivateKey {
+        &self.device_private_key
+    }
+    pub fn device_id(&self) -> &DeviceId {
+        &self.device_id
+    }
+    pub fn name(&self) -> Option<&DeviceName> {
+        self.name.as_ref()
+    }
+    pub fn created(&self) -> &DateTime<Utc> {
+        &self.created
+    }
+    pub fn last_updated(&self) -> &DateTime<Utc> {
+        &self.last_updated
+    }
+}
+
+impl From<DeviceAddResult> for DeviceContext {
+    fn from(dar: DeviceAddResult) -> Self {
+        DeviceContext::new(
+            dar.account_id,
+            dar.segment_id,
+            dar.device_private_key,
+            dar.signing_private_key,
+        )
+    }
 }
 
 /// Generate a device key for the user specified in the JWT.
